requirements.txt
adium-theme-ubuntu==0.3.4
astroid==1.5.3
backports.functools-lru-cache==1.4
click==6.7
configparser==3.5.0
cryptography==1.5
cycler==0.10.0
decorator==4.0.6
enum34==1.1.6
Flask==0.12.2
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
idna==2.1
ipaddress==1.0.16
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==1.5.2rc2
mccabe==0.6.1
mysql-connector-python==2.1.3
MySQL-python==1.2.5
mysql-utilities==1.6.3
numpy==1.11.1rc1
paramiko==2.0.0
pexpect==4.2.0
Pillow==3.3.1
ptyprocess==0.5.1
pyasn1==0.1.9
PyAudio==0.2.11
PyDejavu==0.1.3
pydub==0.20.0
pylint==1.7.4
PyMySQL==0.7.11
pyodbc==3.0.10
pyparsing==2.1.8
pysqlite==2.7.0
python-dateutil==2.4.2
pytz==2014.10
scipy==0.17.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.14
ubuntuone-client-data==14.4
unity-lens-photos==1.0
virtualenv==15.1.0
Werkzeug==0.12.2
wrapt==1.10.11
